+++ test/mutagen_ex_test.exs
@@ -1,8 +1,0 @@
-defmodule MutagenExTest do
-  use ExUnit.Case
-  doctest MutagenEx
-
-  test "greets the world" do
-    assert MutagenEx.hello() == :world
-  end
-end